requirements.txt
pysmb==1.1.27
beautifulsoup4==4.7.1
requests==2.18.4
telepot==12.7